move `internals/structured-clone` content to `modules/web.structured-clone`
